se realiza cambio en la funcion funcioningreso() cambiando el stoi a stof ya que al ingresar datos con decimales este no los tomaba en cuenta, tambien se realiza cambios en las funciones leydeohm() y resistencia del conductor() revisando detalladamente problemas en algunas lineas de codigo y el texto a a mostrar.

+++ funcioningreso.cpp
@@ -35,7 +35,7 @@ string ingreso;
 while(true){
     cout<<"---> "; getline(cin,ingreso);
     try{
-        retorno = stoi(ingreso);
+        retorno = stof(ingreso);
         break;
     }
     catch( invalid_argument& e){
